# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PlantOps Digital Twin Dashboard APIs",
    "description": "Automotive plant digital twin APIs with IBM watsonx Orchestrate integration for energy analysis and predictive maintenance",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://pink-fish-production.up.railway.app",
      "description": "Production Server"
    },
    {
      "url": "http://127.0.0.1:8000",
      "description": "Local Development Server"
    }
  ],
  "paths": {
    "/api/ml/anomaly-detection": {
      "post": {
        "tags": [
          "Machine Learning"
        ],
        "summary": "Detect Anomalies Endpoint",
        "description": "Detect anomalies in plant zone data using Isolation Forest ML model\nReturns anomaly predictions with confidence scores",
        "operationId": "detect_anomalies_endpoint_api_ml_anomaly_detection_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnomalyDetectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnomalyDetectionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/ml/energy-forecast": {
      "post": {
        "tags": [
          "Machine Learning"
        ],
        "summary": "Forecast Energy Endpoint",
        "description": "Forecast energy consumption for a specific zone using trained ML model\nPredicts future energy usage, costs, and CO2 emissions",
        "operationId": "forecast_energy_endpoint_api_ml_energy_forecast_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnergyForecastRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnergyForecastResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AnomalyDetectionRequest": {
        "properties": {
          "zone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Zone",
            "description": "Specific zone to analyze (e.g., 'Paint Shop'), or None for all zones",
            "examples": [
              "Paint Shop",
              "Body Shop (BIW)",
              "Assembly"
            ]
          },
          "hours": {
            "type": "integer",
            "maximum": 168.0,
            "minimum": 1.0,
            "title": "Hours",
            "description": "Hours of historical data to analyze (1-168)",
            "default": 24
          }
        },
        "type": "object",
        "title": "AnomalyDetectionRequest",
        "examples": [
          {
            "hours": 24,
            "zone": "Paint Shop"
          },
          {
            "hours": 48
          }
        ]
      },
      "AnomalyDetectionResponse": {
        "properties": {
          "total_records": {
            "type": "integer",
            "title": "Total Records"
          },
          "anomalies_detected": {
            "type": "integer",
            "title": "Anomalies Detected"
          },
          "anomaly_rate": {
            "type": "number",
            "title": "Anomaly Rate"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/AnomalyResult"
            },
            "type": "array",
            "title": "Results"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "total_records",
          "anomalies_detected",
          "anomaly_rate",
          "results",
          "timestamp"
        ],
        "title": "AnomalyDetectionResponse"
      },
      "AnomalyResult": {
        "properties": {
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          },
          "zone": {
            "type": "string",
            "title": "Zone"
          },
          "is_anomaly": {
            "type": "boolean",
            "title": "Is Anomaly"
          },
          "anomaly_score": {
            "type": "number",
            "title": "Anomaly Score"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "metrics": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Metrics"
          }
        },
        "type": "object",
        "required": [
          "timestamp",
          "zone",
          "is_anomaly",
          "anomaly_score",
          "confidence",
          "metrics"
        ],
        "title": "AnomalyResult"
      },
...
      "EnergyForecastRequest": {
        "properties": {
          "zone": {
            "type": "string",
            "title": "Zone",
            "description": "Zone name to forecast energy consumption",
            "examples": [
              "Paint Shop",
              "Body Shop (BIW)",
              "General Assembly"
            ]
          },
          "hours_ahead": {
            "type": "integer",
            "maximum": 168.0,
            "minimum": 1.0,
            "title": "Hours Ahead",
            "description": "Number of hours to forecast ahead (1-168)",
            "default": 24
          },
          "current_temp": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Temp",
            "description": "Current temperature in \u00b0C (optional, uses latest if not provided)"
          },
          "current_efficiency": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Efficiency",
            "description": "Current efficiency percentage (optional, uses latest if not provided)"
          }
        },
        "type": "object",
        "required": [
          "zone"
        ],
        "title": "EnergyForecastRequest",
        "examples": [
          {
            "current_efficiency": 85.0,
            "current_temp": 185.0,
            "hours_ahead": 24,
            "zone": "Paint Shop"
          },
          {
            "hours_ahead": 48,
            "zone": "Assembly"
          }
        ]
      },
      "EnergyForecastResponse": {
        "properties": {
          "zone": {
            "type": "string",
            "title": "Zone"
          },
          "forecast_hours": {
            "type": "integer",
            "title": "Forecast Hours"
          },
          "forecast": {
            "items": {
              "$ref": "#/components/schemas/ForecastPoint"
            },
            "type": "array",
            "title": "Forecast"
          },
          "total_predicted_energy": {
            "type": "number",
            "title": "Total Predicted Energy"
          },
          "total_predicted_cost": {
            "type": "number",
            "title": "Total Predicted Cost"
          },
          "total_predicted_co2": {
            "type": "number",
            "title": "Total Predicted Co2"
          },
          "confidence_level": {
            "type": "string",
            "title": "Confidence Level"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "zone",
          "forecast_hours",
          "forecast",
          "total_predicted_energy",
          "total_predicted_cost",
          "total_predicted_co2",
          "confidence_level",
          "timestamp"
        ],
        "title": "EnergyForecastResponse"
      },
      "ForecastPoint": {
        "properties": {
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          },
          "hour": {
            "type": "integer",
            "title": "Hour"
          },
          "predicted_energy_kwh": {
            "type": "number",
            "title": "Predicted Energy Kwh"
          },
          "predicted_cost_usd": {
            "type": "number",
            "title": "Predicted Cost Usd"
          },
          "predicted_co2_kg": {
            "type": "number",
            "title": "Predicted Co2 Kg"
          }
        },
        "type": "object",
        "required": [
          "timestamp",
          "hour",
          "predicted_energy_kwh",
          "predicted_cost_usd",
          "predicted_co2_kg"
        ],
        "title": "ForecastPoint"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}